DiscordChessBot/pyproject.toml
[project]
name = "chess-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.11.18",
    "aiosignal>=1.3.2",
    "attrs>=25.3.0",
    "audioop-lts>=0.2.1",
    "chess>=1.11.2",
    "discord-py>=2.5.2",
    "frozenlist>=1.6.0",
    "idna>=3.10",
    "multidict>=6.4.4",
    "propcache>=0.3.1",
    "python-dotenv>=1.1.0",
    "yarl>=1.20.0"
]